Progress check-in of the code for a file-based implementation of the RealmConnector interface (which is proposed to replace the existing RequestSecurityProvider interface in org.apache.tomcat.core).  This code compiles cleanly, but has not been tested.  In fact, there are many XXX comments marking TODO items that must be resolved before it *can* function, and we will need to have a utility program to create the XML file with the encrypted password strings.  However, it is worth publishing this in its current state for review and comment by the community.
There are no linkages to this package in the existing code base yet, so
nothing should get broken.

The included files (package org.apache.tomcat.security.file) are as
follows:

Constants.java -- Application constants for this package, following
	the usual Tomcat pattern (but with at least a couple comments :-).

FileRealmConnector.java -- Implementation of the RealmConnector
	interface, interacting with a FileRealmDatabase object that
	will have been loaded from an XML file.  Note that, as with
	the RealmConnector interface itself, the available methods
	only answer the questions that a Context needs to ask -- in
	particular, the deployment descriptor only knows about users
	(principals) and roles, not about groups.  It is up to each
	RealmConnector implementation to adapt these questions to the
	underlying data structures of the security domain being accessed.

FileRealmDatabase.java -- Database object representing the contents
	of an XML file conforming to the "tomcat-users" DTD.  This is
	a separate object because it will be needed by administrative
	applications that create and update users, groups, and roles.

FileRealmGroup.java -- In-memory representation of a group, with its
	associated users and roles.

FileRealmUser.java -- In-memory representation of a user, with its
	associated groups and roles.

tomcat-users.dtd -- DTD for a file-based security domain (this is
	basically Pier's proposed DTD with some comments added)

tomcat-users.xml -- Example "database" contents that conforms to the
	DTD (this is basically the example proposed on the TOMCAT-DEV
	mailing list).

Once the RealmConnector concept is accepted and integrated with the core,
additional implementations for other security domains need to be
developed.  In particular, a RealmConnector implementation that utilizes
the Apache connector's linkage back to Apache is needed for installations
of the Apache + Tomcat combination.