Removed RequestImpl, Request is back to its original form.
The interface to be implemented by Server Adapters is
RequestAdapter.
Changing Request to an interface didn't worked - Request
is used internaly in many places, so removing methods
is too difficult, and in general, Request serve a different
purpose and needs all the extra methods.

RequestAdapter will be "pluged-in" as soon as I finish
converting existing adapters to implement the new interface.