VM crashes when loading classes from a persistent workdir
The error is caused when the class name is not set in the JSPEngineContext.
The call to the classLoader's findLoadedClass method was being passed null.
The fact that this causes a SIGSEGV/Access violation with newer VMs is not a
good thing.

The reason the class name being passed is null comes from two problems.
Firstly in org.apache.jasper.compiler.Compiler, the compile method will
return if the class file is not outdated before setting the class' name from
the mangler. Secondly the mangler, org.apache.jasper.compiler.JspCompiler,
is not setting the class name if the class file was not outdated.

The following patches just move code around to ensure the class name is set
even when the class file is not outdated. There is also a patch for a
cosmetic error.
Submitted by: Conor MacNeill <conor@m64.com>