This is the patch that enables use of the new Tomcat.Next session management implementation.  If you need to return to the original implementation, simply reverse this patch.
Once we are satisfied that the new session management code works,
we can refactor out the extra "SessionManager" layer.  The part of a
session manager that knows about the semantics of session management
should be returned to the core package, leaving the "Manager"
implementation as an intelligent database of active sessions.