One step towards Craig's design for sessions.
The session manager no longer needs to process the request, it
acts just as a repository.
The request processing ( extracting the session Id) is done by
an interceptor. The other way - setting the response after a new
session is created should also be in a post-service interceptor.
( this way you can configure SSL sessions or URL rewriting without
touching DbManager )