Move DefaultServlets, InvokerServlet and NoInvokerServlet in a separate directory.
The idea is to keep in core only the classes and interfaces that are required, anything
that is "plugeable" will be in a separate directory.

This will help people to work on different sub-projects, and to understand what
happens, and will also allow us to simplify and document the interfaces between
components.