- few name changes ( lookupResult -> subRequest, server-> contextM ) - removed ForwardedRequest - it was used only in forwardRequest and with without LookupResult ( i.e. with sub-Request) it is useless.
- started the new subRequest mechanism. This is the last big change, and it may
brake some code.
This will affect:
- Context.getResource()
- Context.getContext()
- Context.getRealPath()
- RequestDispatcher ( forward and include )
( and few other places ).

How it works:
There are few kinds of subRequests:
- RequestDispatcher will generate a new request and execute it
- few methods need to go through the request processing, but without executing.
For example getContext() needs to call the first set of interceptors that sets the
context. getRealPath() and getResource() - same, plus all the URLMappings

The main entry point for all this is ContextManager, and the mechanism for a
subRequest is identical/shared with the normal request processing done for
Adapters.


PR:
Obtained from:
Submitted by:
Reviewed by: