Moved back to the original model of RequestDispatcher.
The most I could do is to document it and make it look
better (IMHO) - it is still the most difficult part of
tomcat ( and of the spec ).

Please - in case you want to start from scratch with
a new architecture, do start with RequestDispatcher and
an API where RD will fit - all the rest is trivial.