Last big change before freeze.
Removed one extra layer - Request/ResponseAdapter, which duplicated
the Request interface.

Now all that remains to do is clean up Request/Response interfaces -
it contains all methods that are used ( or were used !) by tomcat,
in order to support the "evolution". Probably both can be reduced
in half.

The reason RequestAdapter was added is that Request was far too
complex and had many hardcoded deps to HTTP and tomcat.server.
Both are no longer valid - the original http server was replaced
with a normal adapter, and Request is going to be a minimal
interface ( like in Craig's proposal, maybe plus few other methods).

There are 2 tests affected by this move - I'm working on a fix.