Converted all tomcat tests to use <gtest> ( golden-file-based-test ).
Probably it should be moved in a new directory ( with the golden
results from share/test/jsp/Golden ) - and the build should
use the new directory.

If nobody -1 this change, we can remove all moo deps from
tomcat build and moo will remain part of jakarta-watchdog.

Note that GTest.java is just a 1 1/2-day hack, it can be cleaned
up and improved ( too many attributes, better support for
checking headers in result, etc), but it should be easier than
changing 25 java files.