Another big change - use the new Tomcat starter ( still using server.xml ) This change is required in order to allow adding of new "modules" ( interceptors, adapters, etc) at config time, and uses ant-like "patterns" to configure independent components. It is not final, but it's better to do the move now in order to "burn" it and to allow development of adapters. ( and also get feedback on the "hierarchy" patterns - which is an extension to ant )
Note that stoping tomcat is done using one of the existing adapters
( Ajp12), which has a "stop" command inside (used by Apache to
restart jserv). We will add RMI later,
after more admin methods will be added ( that also mean you need
to start ajp12 instead of RMI !). It no longer use a log file,
but you need to specify the same server.xml when you start and
stop.

Again - it probably brake something, but that's good because we'll
get a chance to fix it, and we can discover more things to do!