Big changes in Response.java - removed all "extra" methods, removed most of the XXX code in ResponseImpl.
Craig, please take a loog at Response, we should try to use few common
interfaces between Catalina and main - most of the methods in response are
needed (IMHO), and probably you'll need to add them when you'll be close
to a working Catalina to support full 2.2 ( buffering, etc) and for connectors.

As part of the changes, the fixHeaders() moved to an interceptor and will be
customizable ( since Apache sets it's own Server and Date header) and "custom"
code can add special headers. The "beforeBody" method is used to give
interceptors the chance to alter headers/state before being sent to user,
and will always  be called before the first write to the Writer.

All error processing code is now in DefaultErrorPage servlet ( used to be
in ResponseImpl and ResponseWrapper). It's still hardcoded, but it's easy to
replace it ( to add better/different reporting styles).

Also changed the names of RequestInterceptor methods to reflect what happens
at each stage - still need to figure an efficient way to set interceptors only
for a number of "stages". Still need to add authenticate() and authorize()
methods ( similar with Apache, IIS, NSAPI )