- Changed the names of handler methods in request and context interceptors ( removed the handle* prefix )
- Added a base class for request/context interceptors

- Major change in loadOnStartup - Context is no longer involved, all the code
is in LoadOnStartupInterceptor. The "loadOnStartup" property is part of
ServletWrapper, instead of keeping a hashtable mapping SW to loadOnStartup.

- Fixed Request.getPathTranslated() - it used to return the path for the current request instead
of the translated pathInfo

- removed read() and write() methods from Request/Response - they belong to Stream.
Request/ResponseImpl still have the methods (overriden by adapters), in order to simplify
adapters ( they just have to extend Request/ResponseImpl, no need to deal with Streams if
they don't need to)

- getRealPath() - code was redundant, and no need for URLs ( since real path works only
for "normal" file-system based contexts)

- removed getServletsByPath(), and fixed containsJSP ( the only method using it) to
return when the first JSP is found.

- removed getServletsByClassName. It was a private and un-used method. Can be implemented
outside context if anyone needs it.

- removed empty ( and unused ) methods from ContextManager

- removed Request.getParametersCopy(), replaced it with equivalent code in RDispatcher

- removed ResolvedServlet from Request - not used anywhere, can be added if we find a user
( and if we find what it does :-). Same for ResourceName. Same for ServletBase

- Added few more comments ( surprise !)

- Fixed HttpResponseAdapter to send multiple headers.

- Changed tomcat.sh to load all jars in tomcat/lib

Sorry for the number/length of the changes, but I was offline and most
of them are related ( cleanup, reorg ). The target is to document/clean Request, Response,
Context and ContextManager.