- Moved the JSP "load-on-startup" hack in LoadOnStartupInterceptor. Still a hack, but no longer in core.
- As a result we can remove JSPWrapper. We still have a special case
for JSP pages, but it will be easy to fix ( i.e. have JSP engine set
the generated servlet wrapper after the first invocation - or another
way to avoid a RD )

- Fixed getPathTranslated() ( regression after cleanup )

- added "deprecated" comments in Context, many methods will go

- added few more debug lines, will probably be removed for release
or replaced with if(debug). Time the initialization ( the new web.xml
and server.xml reader will be 2x faster )

- servlet initParam are set on ServletWrapper instead of Context.