- Added a new ServletLoader interface, with API support for reloading. - Added a context property "reloadable" - Added a SerivletLoader implementation that plugs in AdaptiveClassLoader
Please note that this is just code to clarify core interfaces, not
an implementation for class reloading ( which require a lot more).

( right now class reloading should work - i.e. if you change a class
it will reload it, but the internal state will be a mess. I have no ideea
how it will interact with JSP reloading either.)