Ok, I hope this is the last "big" change for 3.1
Error handling was broken ( i.e. error-page was not called for error codes,
exceptions were handled at-hoc).

Added a single method that handles that, in ContextManager, added code to
redirect to error pages for special HTTP errors, try hard to avoid loops.

We have a bit too much debugging on, but that's not a but and will be removed
before 3.1 ( since we are bug-fixing we should keep it until everything
works )

Implemented form login, added an example. Now we should have BASIC and FORM
working ( except that no auth repository is plugged in, any user/password will
be accepted ). I'll add users.xml before 3.1 ( as a bug-fix, not a feature :-)