BUGFIX: fixed various bugs in error processing, security checking, form-based login.
ADDED:
- simple user database ( tomcat-users.xml ) - it's very limited, but should
work for standalone tomcat - and we want to plug an existing powerfull auth,
not to design one ( based on Craig's security code, I simplified it a bit to
be able to get it done )

- error handler will log  the stack traces.

- few more messages and traces if debug is enabled.

- Removed an obvious hole - after authentication we'll save a private object
in the session, instead of a "String username". The only way you can create the
Credential is via SecurityCheck, that means a trojan servlet can't influence
another application.

Also: tested form login, it seems to work - fill bugs in
jakarta.apache.org/bugs !!!