Few more cosmetic changes, added few debug lines.
It seems we pass all but 3 watchdog tests ( and I'm not sure it's a problem
with tomcat or with the tests).
We have 6 tomcat-test bugs ( all related with error reporting ).

We have some VM-specific problems, it some cases JDK1.2.2/Linux will hang
up ( refuse to server connections ) - I'm unable to start the debuger,
need a bit of time.
We also have a 1.1.8 problem ( IBM vm/Linux ) - I don't know how the
local address is read as 1.0.0.127 instead of localhost or at least
127.0.0.1.