Verified tomcat on Win98 working standalone, with IIS with AJP and JNI connectors. Terific job Gal, the JNI is just great.
Fixed minor problems with directories - make sure the loggers are created
under tomcat home, same for workdir.

Added few more comments and few more checks - Tomcat+JNI+IIS is a very good
test case, no control on start dir or env.
Also made few changes in the IIS, Netsape and JNI documentations and moved
the sample config files into src/etc - it will now be built into tomcat.

The current behavior is a bit more permisive with config problems:

- if -Dtomcat.home is set, it will be used as tomcat installation dir, and
will be used for classpath and default files ( dtds, configs ).
If <contextManager home=""> is specified it will be used as base for webapps,
config, logs, workdir.

- if no tomcat.home is set, we use  contextManager.home as default - it is
also possible to set tomcatHome as an attribute of <contextManager>

- if none is set up, current working directory is used.( we should throw
an exception probably, but give it a chance)