Partial fix for #139.
We need to check the native code too. javax.servlet.http.Cookie needs
to be checked and fixed - but I don't have enough karma :-)

The problem: RFC 2109 and few browsers that are following the standards
( if you can believe that) are sending the cookie value in quotes, and
tomcat is not striping the quotes so the returned value is different from
what the user set.

Still more work needs to be done to check all the code and be correct
Thanks to  Miles Sabin <msabin@cromwellmedia.co.uk> for finding and
kbayer@hms.harvard.edu for fixing the bug.

Submitted by:	kbayer@hms.harvard.edu