Allow a new session to be created after invalidating the previous one, in the context of processing the same request (bug report 55).  This makes the following code sequence work:
	HttpSession session = request.getSession(false);
	if (session != null)
	    session.invalidate();
	session = request.getSession(true);
	long creationTime = session.getCreationTime();

without throwing an IllegalStateException on the last statement.  NOTE:  this
sequence will be ineffective if the response has already been committed (because
you can no longer add the session ID cookie to the response headers), but that
is also true of creating a new session in the first place -- so I don't see
this as a big problem.

PR:55
Submitted by:	jcpage@corp.home.net (bug report), gaborliptak@usa.net (patch)