solved (at least I think) the double initialization (Unix) due to Apache defensive startup. We now start the JVM  only in the 2nd pass (in automatic mode).
This works well if you take care to set the :
ApJServVMTimeout value in jserv.conf big enough to start & initialize the servlet engine + servlets loaded at startup.
This value is for me 10 seconds to start Tomcat, JServ works with 4.

This was I think the major reason to not use automatic startup of Tomcat by Apache

Jean-Luc