Correct the implementation of encodeURL() and encodeRedirectURL() so that they correctly determine when encoding should take place.  For the record, those rules are as follows: - We are in a valid session - The session ID was not requested with a cookie - The URL to be encoded references a URL within the current   web application (that is, we match on the scheme, host, port,   and context path of the absolute URL that corresponds to   the specified URL if it is relative)
In addition, sendRedirectURL() now absolutizes the specified location,
even though this is already done in the default error handling servlet,
so that the spec rule (that it must be absolutized) will be obeyed even
if the web app has defined their own error page for SC_MOVED_TEMPORARILY
(legal but not likely).

Tested with URLs that are absolute, host relative (that is, starting with
a slash), and request-relative (not starting with a slash).  Also passes
the current Watchdog test suite.
PR:174
Submitted by:	evan@netsco.com