Correct a spec compliance bug in the way extension mapping was implemented.  Previously, a request for
	http://localhost:8080/examples/jsp/snoop.jsp/foo

was being resolved to the /examples/jsp/snoop.jsp JSP page, and passing a
path info value of "/foo".  This is contrary to the servlet specification
version 2.2, section 10.1, bullet point 3:

	If the last node of the url-path contains an
	extension (.jsp for example), the servlet
	container will try to match a servlet that handles
	requests for that extension.  An extension is
	defined as the part of the path after the last
	'.' character.

Now, the request shown above will return a 404 error (file not found).

This also required fixing a bug in DefaultFileServlet that would throw an
IOException ("bad file path") on the above request, because
getCanonicalPath() choked when it tried to treat the "snoop.jsp" part as a
directory, when it was actually a file.

Also fixed an additional bug in DefaultFileServlet, which was gluing
together the document root (with a '/' appended) and the path information
from the request, resulting in two slashes in a row.  Apparently most OS's
ignored this (or the underlying JVM fixed it transparently), but it's
better to be correct.

Watchdog runs 100% against this revised code.