Fix a bug where BASIC authentication fails to trigger the pop-up dialog box for username and password on IE.  This bug was introduced by my previous fix to HttpServletResponseFacade.java to make it reset the response inside sendError() and sendRedirect().  Unfortunately, due to the way Tomcat is currently structured, this caused the "WWW-Authenticate" message containing the challenge to be erased.
Netscape Navigator saw the 401 (Unauthorized) error, and popped up a
dialog box anyway, with an "unknown" realm.  However, IE didn't see
a "WWW-Authenticate" so it didn't do anything.

The workaround in this patch is to NOT reset the response if sendError
is called with a status code of 401.  I've been staring at this code for
three hours, and cannot see any other way short of a major restructuring
to get around this problem differently.