Modify the ObjectCreate class to operate in the manner that (I believe) was originally intended, based on which constructor you call.
If the ObjectCreate(classN) constructor is called, an instance of the
class specified by "classN" will always be instantiated.

If the ObjectCreate(classN,attrib) constructor is called, the "classN"
argument is considered to be the default class name.  If an attribute of
the name specified by the "attrib" argument is present in the actual XML
node being parsed, its value is considered to override the default class
name.  (Previously, this triggered a null pointer exception if the
attribute was missing).

This approach is very useful in dynamically configuring new objects, but
allowing the provision of a default class name that can be overridden
instead of requiring it to be explicitly included on *every* XML element.