- Removed the fix for bug #116 ( class loader problems with JNDI and other libs)
- Added back the fix in a better form ( plus support for destroy ). It is now
possible to use it only if you need to ( i.e. no need to do several
introspections if you are in jdk1.1). We use a normal interceptor that will
set the contextClassLoader for JDK1.2 - the code is still jdk independent,
but probably it's usefull to just call the real method instead of using
invoke() - at least faster...

- Start the missing piece for Facade - the ability to restrict access to
tomcat internals. It's still a lot of work in this area - review all
public methods and constructors, including attributes that are set during
jsp processing.