- Added SimpleHashtable from xml-contrib ( Crimson ) - it's a non-synch Hashtable that will improve the performance of the path mapper
- Added "notes" as a general communication mechanism for interceptors.
Notes are similar with req->notes in apache ( I don't remember the
equivalent in IIS, NES).
Remember that interceptors are supposed to be state-less and have as little
config as possible - the state and config is part of the tomcat.core.
That also address Craig's concerns regarding state in interceptors.

- Added mapType property in Container. Every container represents a mapping,
and this allow easy generalization if new mapping types are introduced.

- Every context will also have a container, coresponding to the default
mapping. That will eliminate a lot of checking and guessing, and allows
us to treat all URLs in an uniform way. Note that defaultServlet and a number
of methods are no longer needed, and Context will be simplified a lot.

- Removed the use of request.lookupPath and replaced it with
ServletPath+PathInfo. It was an artificial construct used in the mapper.

- in HttpRequest, lazy evaluation for expensive getHostName()

- fix Invoker ( register the right prefix mapping for further invocations )

I will also check in a cleaned-up version of SimpleMapper ( really cleaned up -
probably 1/3 hashtable lookups and finally clear code ). I will also add
virtual host, but I have to do it from home.