Initial check in for the "next" mapper. It's not yet final, but I cleaned up most of the alghoritm and factored out the prefix matching alghoritm, while removing a lot of redundant operations.
Missing:
- virtual host ( easy, but I want to do it in the right way, by using a
hierarchy mapping instead of simple string hashtable lookups)

- security - as you know, the implementation of security was mostly
experimental, with double a matching - once for the normal request
and once for security checking. This can ( and should ) be done
in one step only, and requires merging the security and mapping patterns.

- tree instead of hashtable, better longest-match ( now we remove
the last component from the path till it matches and look in a hashtable).
A better way is to look from start and stop when no longer matches, or
even better to use a sorted tree. Even with the current simple matching,
using a TreeMap adds 20% to the speed.
Another important thing is to use char[] instead of string operations
when matching - there is no need to create so many strings in the process.

- shortcut code. We do have an experimental "mapCache" that will benefit from
the "locality" ( a number of pages tend to be very popular and frequently
accessed, and we can avoid the full matching alghoritm for those), but we
need to better integrate with the Web adapters.

The code still has a bit of "inheritance" from the (ugly) SimpleMapper,
but it will go away soon.